Create Tree
Using the same basic idea from the C++ code, I used a list to create the tree spaces so that it would be properly spaced. Python doesn't seem to have switch-case statements and would need a class to implemented to have the correct use of the switch-case. But by implementing the list as such the switch-case statement is essentially the same thing.